move round (which is formatting) to calculated
